--- apps/codebility/app/home/applicants/_components/ApplicantsList.tsx
@@ -1,33 +1,0 @@
-"use client";
-
-import ApplicantsTableContainer from "@/app/home/applicants/_components/ApplicantsTableContainer";
-import { ApplicantStatus, Codev } from "@/types/home/codev";
-
-interface ApplicantsListProps {
-  applicants: Codev[];
-  trackAssessmentSent?: (applicantId: string) => void;
-  sentAssessments?: Record<string, boolean>;
-  activeTab: ApplicantStatus;
-  onStatusChange?: () => void; // Add onStatusChange prop
-}
-
-const ApplicantsList = ({
-  applicants,
-  trackAssessmentSent,
-  sentAssessments,
-  activeTab,
-  onStatusChange,
-}: ApplicantsListProps) => {
-  // All filtering and sorting is now handled in the parent component
-  return (
-    <ApplicantsTableContainer
-      applicants={applicants}
-      trackAssessmentSent={trackAssessmentSent}
-      sentAssessments={sentAssessments}
-      activeTab={activeTab}
-      onStatusChange={onStatusChange} // Pass onStatusChange to container
-    />
-  );
-};
-
-export default ApplicantsList;
