+++ apps/codebility/app/home/applicants/_components/_table/applicantMobileTable.tsx
