+++ apps/codebility/app/home/applicants/_components/_table/applicantProfileColSec.tsx
