+++ apps/codebility/app/home/applicants/_components/_table/appplicantColumns.tsx
@@ -1,10 +1,8 @@
 "use client";
 
-import React, { useEffect, useMemo, useState } from "react";
+import React from "react";
 import Image from "next/image";
 import Link from "next/link";
-import { getTestDate } from "@/app/applicant/waiting/_service/util";
-import { getReApplyDate } from "@/app/auth/declined/_service/util";
 import { Button } from "@/Components/ui/button";
 import { IconGithub, IconLink } from "@/public/assets/svgs";
 import { ColumnDef } from "@tanstack/react-table";
@@ -13,6 +11,9 @@ import { ArrowUpDown } from "lucide-react";
 import { Checkbox } from "@codevs/ui/checkbox";
 
 import { NewApplicantType } from "../../_service/types";
+import ApplicantReapplyTime from "../applicantReapplyTime";
+import ApplicantTechStack from "../applicantTechStack";
+import ApplicantTestTimeRemaining from "../applicantTestTimeRemaining";
 import ApplicantActionButton from "./applicantActionButton";
 import ApplicantProfileColSec from "./applicantProfileColSec";
 
@@ -199,47 +200,11 @@ export const applicantsColumns: ColumnDef<NewApplicantType>[] = [
     },
     cell: ({ row }) => {
       const applicant = row.original;
-      const [showAll, setShowAll] = React.useState(false);
-
-      const displayStacks = showAll
-        ? applicant.tech_stacks
-        : applicant.tech_stacks?.slice(0, 5);
-
-      const hasMoreStacks = applicant.tech_stacks?.length > 5;
 
       return (
-        <div className="px-3 py-3 text-sm">
-          <div className="flex w-40 flex-wrap justify-center gap-2 text-sm">
-            {applicant.tech_stacks?.length > 0 &&
-            !applicant.tech_stacks.includes("none") ? (
-              <div className="flex flex-wrap gap-2 text-sm">
-                {displayStacks.map((stack) => (
-                  <Image
-                    key={stack}
-                    src={`/assets/svgs/icon-${stack.toLowerCase()}.svg`}
-                    alt={stack}
-                    width={22}
-                    height={22}
-                    className="h-6 w-6"
-                  />
-                ))}
-                {hasMoreStacks && (
-                  <button
-                    /*  variant="ghost" */
-                    onClick={() => setShowAll(!showAll)}
-                    className="text-sm text-gray-400 hover:text-gray-200"
-                  >
-                    {showAll
-                      ? "Show less"
-                      : `+${applicant.tech_stacks.length - 5} more`}
-                  </button>
-                )}
-              </div>
-            ) : (
-              <span className="text-sm text-gray-500">None</span>
-            )}
-          </div>
-        </div>
+        <>
+          <ApplicantTechStack applicant={applicant} />
+        </>
       );
     },
     meta: {
@@ -321,77 +286,11 @@ export const applicantsColumns: ColumnDef<NewApplicantType>[] = [
     },
     cell: ({ row }) => {
       const applicant = row.original;
-      const applicantData = applicant.applicant;
-
-      const reapplyDate = useMemo(
-        () =>
-          getTestDate(new Date(applicantData?.test_taken || "") || new Date()),
-        [applicantData?.test_taken],
-      );
-
-      const [timeLeft, setTimeLeft] = useState(() => {
-        const now = new Date();
-        const difference = reapplyDate.getTime() - now.getTime();
-        const isExpired = difference <= 0;
-
-        if (isExpired) {
-          return { days: 0, hours: 0, minutes: 0, seconds: 0, isExpired: true };
-        }
-
-        const days = Math.floor(difference / (1000 * 60 * 60 * 24));
-        const hours = Math.floor((difference / (1000 * 60 * 60)) % 24);
-        const minutes = Math.floor((difference / 1000 / 60) % 60);
-        const seconds = Math.floor((difference / 1000) % 60);
-
-        return { days, hours, minutes, seconds, isExpired };
-      });
-
-      useEffect(() => {
-        const interval = setInterval(() => {
-          const now = new Date();
-          const difference = reapplyDate.getTime() - now.getTime();
-
-          if (difference <= 0) {
-            setTimeLeft({
-              days: 0,
-              hours: 0,
-              minutes: 0,
-              seconds: 0,
-              isExpired: true,
-            });
-            clearInterval(interval);
-            return;
-          }
-
-          const days = Math.floor(difference / (1000 * 60 * 60 * 24));
-          const hours = Math.floor((difference / (1000 * 60 * 60)) % 24);
-          const minutes = Math.floor((difference / 1000 / 60) % 60);
-          const seconds = Math.floor((difference / 1000) % 60);
-
-          setTimeLeft({ days, hours, minutes, seconds, isExpired: false });
-        }, 1000);
-
-        return () => clearInterval(interval);
-      }, [reapplyDate]);
 
       return (
-        <div className="py-3 text-center text-sm text-gray-300">
-          {!timeLeft.isExpired ? (
-            <>
-              {timeLeft.days > 0 && (
-                <span>{`${timeLeft.days}d ${timeLeft.hours}h`}</span>
-              )}
-              {timeLeft.hours === 0 && timeLeft.minutes > 0 && (
-                <span>{`${timeLeft.minutes}m`}</span>
-              )}
-              {timeLeft.minutes === 0 && timeLeft.seconds > 0 && (
-                <span>{`${timeLeft.seconds}s`}</span>
-              )}
-            </>
-          ) : (
-            <span className="text-sm text-gray-500">Expired</span>
-          )}
-        </div>
+        <>
+          <ApplicantTestTimeRemaining applicant={applicant} />
+        </>
       );
     },
     meta: {
@@ -438,73 +337,11 @@ export const applicantsColumns: ColumnDef<NewApplicantType>[] = [
     },
     cell: ({ row }) => {
       const applicant = row.original;
-      const reapplyDate = useMemo(
-        () =>
-          applicant?.date_applied
-            ? getReApplyDate(new Date(applicant.date_applied))
-            : new Date(),
-        [applicant?.date_applied],
-      );
-
-      const [timeLeft, setTimeLeft] = useState(() => {
-        const now = new Date();
-        const difference = reapplyDate.getTime() - now.getTime();
-        const isExpired = difference <= 0;
-
-        if (isExpired) {
-          return { days: 0, hours: 0, minutes: 0, seconds: 0, isExpired: true };
-        }
-
-        const days = Math.floor(difference / (1000 * 60 * 60 * 24));
-        const hours = Math.floor((difference / (1000 * 60 * 60)) % 24);
-        const minutes = Math.floor((difference / 1000 / 60) % 60);
-        const seconds = Math.floor((difference / 1000) % 60);
-
-        return { days, hours, minutes, seconds, isExpired };
-      });
-
-      useEffect(() => {
-        const interval = setInterval(() => {
-          const now = new Date();
-          const difference = reapplyDate.getTime() - now.getTime();
-
-          if (difference <= 0) {
-            setTimeLeft({
-              days: 0,
-              hours: 0,
-              minutes: 0,
-              seconds: 0,
-              isExpired: true,
-            });
-            clearInterval(interval);
-            return;
-          }
-
-          const days = Math.floor(difference / (1000 * 60 * 60 * 24));
-          const hours = Math.floor((difference / (1000 * 60 * 60)) % 24);
-          const minutes = Math.floor((difference / 1000 / 60) % 60);
-          const seconds = Math.floor((difference / 1000) % 60);
-
-          setTimeLeft({ days, hours, minutes, seconds, isExpired: false });
-        }, 1000);
-
-        return () => clearInterval(interval);
-      }, [reapplyDate]);
 
       return (
-        <div className="py-3 text-center text-sm">
-          {!timeLeft.isExpired ? (
-            <div className="text-gray-300">
-              <div className="flex justify-center space-x-1">
-                <span className="">{timeLeft.days}d</span>
-                <span className="">{timeLeft.hours}h</span>
-                <span className="">{timeLeft.minutes}m</span>
-              </div>
-            </div>
-          ) : (
-            <span className="font-medium text-green-500">Available now</span>
-          )}
-        </div>
+        <>
+          <ApplicantReapplyTime applicant={applicant} />
+        </>
       );
     },
     meta: {
